requirements.txt
amqp==5.1.1
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.24.85
botocore==1.27.85
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.2
django-colorfield==0.7.2
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-storages==1.13.1
django-taggit==3.0.0
django-tinymce==3.5.0
django-treebeard==4.5.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
kombu==5.2.4
l18n==2021.3
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==4.2.0
tablib==3.2.1
telepath==0.3
urllib3==1.26.12
vine==5.0.0
wagtail==4.0.2
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
